requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# OpenAI
openai>=1.3.0

# Database
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Vector Store & Embeddings
chromadb>=0.4.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
